requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
click==8.1.2
colorama==0.4.4
decouple==0.0.7
discord-components==2.1.2
discord.py==1.7.3
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
python-dotenv==0.20.0
tinydb==4.7.0
typing-extensions==4.1.1
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.8.0
